+++ ClassLibrary/Media/IAudioSampleSource.cs
@@ -5,7 +5,7 @@
 namespace SipLib.Media;
 
 /// <summary>
-/// Delegate definition for the AudioSamplesReady event of the IAudioSamplesSource interface.
+/// Delegate definition for the SendAudioSamples event of the IAudioSamplesSource interface.
 /// </summary>
 /// <param name="NewSamples">Block of new 16-bit linear PCM audio samples.</param>
 /// <param name="SampleRate">Sample rate in samples/second of the data in the NewSamples array.</param>
@@ -22,7 +22,7 @@ public interface IAudioSampleSource
     public event AudioSamplesReadyDelegate AudioSamplesReady;
 
     /// <summary>
-    /// Tells the audio sample source to start sending audio samples by firing the AudioSamplesReady event.
+    /// Tells the audio sample source to start sending audio samples by firing the SendAudioSamples event.
     /// </summary>
     public void Start();